exa: Replace TLSF with a custom allocator
The main problem with TLSF is that it embeds the allocation description
into the data pool and hence the the internal data could be corrupted by
a erroneous drawing operation. Let's switch to a custom allocator which
actually should be much more efficient and allows us to implement nifty
features like pool memory defragmentation and perform specific optimizations.