exa: pixmap refrigerator
Refrigerator compresses (freezes) pixmaps data when pixmap is idling for
too long and decompresses (thaws) the data when it is needed again. With this
feature I can open gallery with tens of thousands pictures without a hiccup!

Problems with the low amount of CMA memory are solved because instead of
occupying CMA area until it ends, now a window of 16MB is maintained and
pixmaps are getting swapped in/out dynamically. This means that swapped
out and compressed pixmaps could be moved to the systems swap partition
by kernel.