exa: implement DownloadFromScreen
Use accelerated memcpy() for reading of uncached BO data. Note that EXA
doesn't sync for Upload/DownloadFromScreen, hence we should wait for the
fence. It should be possible to implement custom Get/PutImage that take care
of tracking damaged areas and avoiding the stalls when unnecessary, but seems
that will be much more work which could be done later.