driver: remove dirty region tracking
The dirty-region notification is meant to perform cache flushing
operation by the DRM driver. Tegra DRM doesn't support that operation
because it is not really needed for a write-combined mapping and
DirtyFB IOCTL is a no-op for Tegra. Note that "dirty updates" supposed
to be auto-disabled for DRM drivers that do not need that, but the
disabling wasn't working due to a bug/typo in the removed code.