exa: support 1x1 pixmap texture-reduce optimization
Debug messages revealed that there a lot of compositing operations that
use 1x1 texture with coordinates being wrapped, that is equal to operation
that uses a solid color. This fact allows us to reduce 1x1 textures into
a solid color and thus to use much more optimal shader. The performance
win is very-very significant.