exa: composite-3d: fix memory leak
Initially the idea was to avoid reallocation of attributes buffer
whenever possible. Seems code changed somewhere in-between and that
causes attributes buffer to be never released. We probably will return
to the reallocation-avoidance later on, but for now let's just always
release the buffer after job's submission to fix the bug.