exa: better picture format tracking
The pixmap's drawable data is format-less by default and data receives
meaning only during operations with pictures, which are in fact descriptors
for the drawable data. We need to know what is the actual format of the data
to compress it better and assume that the lastly used picture-descriptor
is actual at the time of compression.