exa: rename "_clamped" shaders to "_pad" for clarity
OpenGL and Xrender have different meanings for texture clamping, in OpenGL
world it means to pad texture's border on sampling outside of the texture,
while Xrender wants to sample black color and it uses "pad" terminology
for the OpenGL's "clamping". Let's use Xrender's terminology for the padding
universally and use the name "clipping" for the case of the black texels
sampling.