exa: composite-3d: switch TEX wrapping to pad-mode for emulated normal-mode
GR3D TEX unit supports only padding for non-pow2 textures, other modes
need to be emulated within shader for the non-pow2 textures. We are emulating
normal-wrap for two opOver shaders, leaving HW mode set to the normal-wrap
mode and it works without a cough on T20, but not on T30 that fails "repeat"
rendercheck tests that use those two opOver shaders. Turned out that turning
off normal-wrapping in HW fixes the problem on T30 and now T30 passes all of
the tests.