exa/display: always zero-fill allocated data
For now on Tegra20 all BO's are allocated using DMA API by the kernel
driver and it always zero's the allocated pages. This zero'ing doesn't
happen on T30 which uses get_pages() by default with enabled IOMMU support,
and allocated pages are not zero'ed which results in a noise-garbage being
displayed until something decides to draw to frambuffer, that's not very
nice. Technically data zero'ing adds some overhead, but it should be
negligible.