dri2: not every create buffer is dedicated to Tegra-DRI
Turned out that DRI buffer could be created for a redirected pixmaps
(CREATE_PIXMAP_USAGE_BACKING_PIXMAP hint). In this case the BO's name
may be unavailable (BO is backed by malloc), hence the creation fails
in this case. This patch fixes occasional DRI display initialization
failure of the grate-tests.