display: support 180 degree display panel rotation
Nexus 7 has a display panel that is physically mounted upside down.
Previously we were rotating display planes within kernel driver, but
now the rotation is moved to the Opentegra driver because I'm looking
at upstreaming the kernel patches and turned out that it will be easier
start without the universal display rotation support.