dri2: add fencing after copying operation
Previously, we fenced every drawing operation in the driver. Now we support
asynchronous drawing and this causes a lot of flickering in applications
that use DRI, like glxgears for example. It happens because UAPI v1 doesn't
support implicit fencing and we haven't converted everything to UAPI v2 yet,
so some explicit fencing still needs to be done today.