exa_mm_pool: improve slow compaction policy
Current policy allows to bypass minimum time limitation between compaction
operations if there is a lot of free space in pools. This policy isn't not
optimal for a very frequent re-allocations, which doesn't happen often in
practice, but happens when running stress tests. We can safely remove the
min-time bypass, so let's do it.