exa/composite_3d: fix flush-wait order for upstream UAPI
The tegra_exa_finalize_3d_state() performs flushing of deferred 2d
operations and then tegra_exa_wait_pixmaps() should wait for the
completion of the operations in a case of upstream UAPI, but that
waiting doesn't happen if functions are invoked in a wrong order.

This fixes drawing glitches using vanilla upstream kernel.