exa/optimizations: add missing cancellation of 3d job
Drawing operation may cancel all deferred operations. The cancellation of
3d jobs was missing for optimized 2d jobs, which results into a skipped
flushing of 2d job if there is at least one 3d jobs in a deferred queue
because next enqueued job was skipping the 2d flushing since it assumed
that previous 3d job in the queue already did it. This fixes flashing
text in v1.26 MATE terminal.